feat: add tests via container images
Test container images are produced, which run simple C/CPP binaries
(linked dynamically to glibc) - should the execution of image fail,
it indicates that the image is not built properly.

Images come in two flavours:
1. testGlibcBinary- which used hand-crafted hello world application,
patched to be working akin to 'vanilla' binaries produced by
the gcc toolchain on popular Linux distros.
3. ubuntuDateutils  - which uses unmodified dateutils binaries, downloaded from
Ubuntu 'lunar' .deb repository which are dynamically linked against
glibc 2.37 (in a manner default for Linux FHS).

Co-authored-by: Artur Stachecki <dev@r2r.sh>