feat: introduce image variant with bash
This changeset introduces a image variant with statically compiled
bash added to the image. The chief reason for introducing this flavor
is to reduce amount of changes that many of open-source Bazel rules
would require to make them compatible with starterkit as execution
platform.

Example: One of the most popular gcc toolchain [1] which deems itself
to be "hermetic", relies on the presence of bash interpreter and
coreutils - without them, it fails.

[1] https://github.com/f0rmiga/gcc-toolchain